+++ src/agent/loaders.rs
@@ -7,10 +7,11 @@ use flate2::write::GzEncoder;
 use flate2::Compression;
 use regex::Regex;
 use secrecy::ExposeSecret;
-
+use serde_json::Value;
 use std::fs::File;
 use std::io::Read;
 use std::io::Write;
+use walkdir::WalkDir;
 
 use chrono::Utc;
 use log::{debug, error, info, warn};
@@ -40,14 +41,14 @@ pub fn use_filesystem() -> bool {
 pub trait FileLoader {
     // utils
     fn build_filepath(&self, doctype: &String, docid: &String) -> Result<PathBuf, Box<dyn Error>>;
+    fn build_file_directory(&self, doctype: &String) -> Result<PathBuf, Box<dyn Error>>;
     fn build_key_filepath(
         &self,
         doctype: &String,
         key_filename: &String,
     ) -> Result<PathBuf, Box<dyn Error>>;
 
-    // init
-    fn fs_docs_load_all(&mut self) -> Result<Vec<String>, Box<dyn Error>>;
+    fn fs_docs_load_all(&mut self) -> Result<Vec<String>, Vec<Box<dyn Error>>>;
     fn fs_agent_load(&self, agentid: &String) -> Result<String, Box<dyn Error>>;
     // fn fs_agent_new(&self, filename: &String) -> Result<String, Box<dyn Error>>;
     // fn fs_document_new(&self, filename: &String) -> Result<String, Box<dyn Error>>;
@@ -89,11 +90,11 @@ pub trait FileLoader {
 
 #[cfg(not(target_arch = "wasm32"))]
 impl FileLoader for Agent {
-    fn build_filepath(&self, doctype: &String, docid: &String) -> Result<PathBuf, Box<dyn Error>> {
+    fn build_file_directory(&self, doctype: &String) -> Result<PathBuf, Box<dyn Error>> {
         if !use_filesystem() {
             let error_message = format!(
-                " build_filepathFilesystem features set to off with JACS_USE_FILESYSTEM: {} {}",
-                doctype, docid
+                " build_file_directory Filesystem features set to off with JACS_USE_FILESYSTEM:  {}",
+                doctype
             );
             error!("{}", error_message);
             return Err(error_message.into());
@@ -103,7 +104,10 @@ impl FileLoader for Agent {
         let jacs_dir = env::var("JACS_DATA_DIRECTORY").expect("JACS_DATA_DIRECTORY");
 
         let path = current_dir.join(jacs_dir).join(doctype);
-
+        return Ok(path);
+    }
+    fn build_filepath(&self, doctype: &String, docid: &String) -> Result<PathBuf, Box<dyn Error>> {
+        let path = self.build_file_directory(doctype)?;
         let filename = if docid.ends_with(".json") {
             docid.to_string()
         } else {
@@ -211,9 +215,37 @@ impl FileLoader for Agent {
         self.set_keys(private_key, public_key, &key_algorithm)
     }
 
-    /// on instantiation load and validata all local documents
-    fn fs_docs_load_all(&mut self) -> Result<Vec<String>, Box<dyn Error>> {
-        Err(not_implemented_error())
+    /// function used to load all documents present
+    fn fs_docs_load_all(&mut self) -> Result<Vec<String>, Vec<Box<dyn Error>>> {
+        let mut errors: Vec<Box<dyn Error>> = Vec::new();
+        let mut documents: Vec<String> = Vec::new();
+        let mut paths: Vec<PathBuf> = Vec::new();
+        let agent_path = self.build_file_directory(&"agent".to_string()).unwrap();
+        paths.push(agent_path);
+        let document_path = self.build_file_directory(&"documents".to_string()).unwrap();
+        paths.push(document_path);
+        for path in paths {
+            // now walk the directory using walkdir crate
+            for entry in WalkDir::new(path)
+                .into_iter()
+                .filter_map(|e| e.ok())
+                .filter(|e| e.path().is_file())
+            {
+                let document_string = fs::read_to_string(entry.path());
+                match document_string {
+                    Ok(doc) => {
+                        documents.push(doc);
+                    }
+                    Err(e) => {
+                        errors.push(Box::new(e));
+                    }
+                }
+            }
+        }
+        if errors.len() > 0 {
+            error!("errors loading documents {:?}", errors);
+        }
+        Ok(documents)
     }
 
     fn fs_agent_load(&self, agentid: &String) -> Result<String, Box<dyn Error>> {